CORE-5679 - refactor unnecessary class for hook registration / handler to use extension function
